+++ packages/components/src/components/common/label/label-styles.ts
@@ -1,6 +1,6 @@
 import { spacingStaticXs, proseTextSmStyle } from '@porsche-design-system/emotion';
 import type { JssStyle, Styles } from 'jss';
-import { getHiddenTextJssStyle, getTransition } from '../../../styles';
+import { getDisabledBaseStyles, getHiddenTextJssStyle, getTransition } from '../../../styles';
 import { colorContrastHigh, colorPrimary } from '../../../styles/css-variables';
 import { buildResponsiveStyles } from '../../../utils';
 import type { BreakpointCustomizable } from '../../../utils/breakpoint-customizable';
@@ -36,6 +36,7 @@ export const getFunctionalComponentLabelStyles = (
       cursor: isDisabledOrLoading ? 'not-allowed' : 'pointer',
       ...(isDisabledOrLoading && {
         pointerEvents: 'none', // prevents label interaction when disabled or loading
+        ...getDisabledBaseStyles(),
       }),
       color: colorPrimary,
       transition: getTransition('color'), // for smooth transitions between e.g. disabled state