+++ packages/components/src/components/multi-select/multi-select/multi-select-styles.ts
@@ -1,10 +1,5 @@
 import { borderWidthBase, spacingStaticXSmall } from '@porsche-design-system/emotion';
-import {
-  addImportantToEachRule,
-  getDisabledBaseStyles,
-  hostHiddenStyles,
-  preventFoucOfNestedElementsStyles,
-} from '../../../styles';
+import { addImportantToEachRule, hostHiddenStyles, preventFoucOfNestedElementsStyles } from '../../../styles';
 import { formElementPaddingHorizontal, getCalculatedFormElementPaddingHorizontal } from '../../../styles/form-styles';
 import {
   getButtonJssStyle,
@@ -52,7 +47,6 @@ export const getComponentCss = (
           [`${cssVarInternalMultiSelectOptionScaling}`]: isCompact ? 0.64285714 : 1,
           [`${cssVarInternalOptgroupScaling}`]: isCompact ? 0.64285714 : 1,
           ...hostHiddenStyles,
-          ...(isDisabled && getDisabledBaseStyles()),
         }),
       },
       ...getFunctionalComponentLabelAfterStyles(),