+++ packages/components/src/components/pin-code/pin-code-styles.ts
@@ -50,7 +50,6 @@ export const getComponentCss = (
         [`${cssVarInternalPinCodeScaling}`]: isCompact ? 0.64285714 : 1,
         ...addImportantToEachRule({
           ...hostHiddenStyles,
-          ...(isDisabled && getDisabledBaseStyles()),
         }),
       },
       ...getFunctionalComponentLabelAfterStyles(),
@@ -73,7 +72,7 @@ export const getComponentCss = (
         textOverflow: 'ellipsis',
         cursor: isDisabled || isLoading ? 'not-allowed' : 'text',
         textAlign: 'center',
-        ...(isLoading && getDisabledBaseStyles()),
+        ...((isDisabled || isLoading) && getDisabledBaseStyles()),
         '&:focus-visible': {
           borderColor: formStateBorderHoverColor,
         },