+++ packages/components/src/components/select/select/select-styles.ts
@@ -1,10 +1,5 @@
 import { borderWidthBase, spacingStaticXSmall } from '@porsche-design-system/emotion';
-import {
-  addImportantToEachRule,
-  getDisabledBaseStyles,
-  hostHiddenStyles,
-  preventFoucOfNestedElementsStyles,
-} from '../../../styles';
+import { addImportantToEachRule, hostHiddenStyles, preventFoucOfNestedElementsStyles } from '../../../styles';
 import { formElementPaddingHorizontal, getCalculatedFormElementPaddingHorizontal } from '../../../styles/form-styles';
 import {
   getButtonImageJssStyle,
@@ -54,7 +49,6 @@ export const getComponentCss = (
           [`${cssVarInternalOptgroupScaling}`]: isCompact ? 0.64285714 : 1,
           ...hostHiddenStyles,
         }),
-        ...(isDisabled && getDisabledBaseStyles()),
       },
       ...getFunctionalComponentLabelAfterStyles(),
       ...preventFoucOfNestedElementsStyles,